Rewrote indirect "adf_mode" code to be more direct
Replaced the clever, yet totally confusing, code using the "adf_mode"
variable.  This variable was used with a complex interaction of bit shifts
and subtraction indirectly influnced by the value of s->source_mode.  The
replacement code direct references s->source_mode instead, simplifying
things greatly.  This change is functionally identical.

As a side note, the adf_mode code allowed "ADF Rear" scanners, in addition to
"ADF Duplex" scanners, to enter the section of code normally used for (HP)
duplex scanners.  It's not clear if this was the intended effect, or if, as
the comments suggest, this was an unnoticed side effect caused by the
confusing code.  The point is moot, however, because prior in the code, the
ADF Duplex flag is set anytime the ADF Rear flag is set.  As long as all
ADF Rear scanners are duplex scanners this shouldn't be a problem.