Return copied values in sane_get_parameters()
Until now the code has been changing the frontend supplied pointer to point
to our internal data structure.  There are two problems with this.  It
orphans the data previously associated with that pointer, which causes a
memory leak.  It also gives the frontend the ability to munge our internal
data structure.

The only downside to this change is the possibility that a frontend took the
spec too literally and merely supplies a pointer but does not allocate
storage space for the result.  Technically, the spec is a bit vague on this
point, because it does not specifically state who should allocate the
structure, but I feel a reasonable programmer would guess that space
allocation is a frontend obligation.