Fixed flipping duplexer initialization
For scanners which potentially have flipping duplexers, only enable
duplexing if there is an ADF present.  This leads to the next problem, that
the flipping duplexers (present only on the HP 8200 series AFAIK) don't
always seem to be properly configured when powered up.  To solve this some
magical values are written to magical memory locations, and some more
magical status is read back, causing an apparent firmware reset which
blocks communication for a short period of time.  See adf_reset() for
details.