Skip post-processing when caching flipping duplex
In a previous patch, I attempted to avoid the possibility of a page remaining
partially inside the ADF unit of a flipping duplex scanner after a scan.
Unfortunately, I introduced a bug where the backside page was accidentally
returned to the client in addition to being cached.  This was because I had
a misunderstanding of the post-processing code.  This patch skips the post-
processing section altogether, avoiding any unnecessary post-processing and
subsequent unwanted output.  The other alternative here was to try to monkey
with the buffers (as is done for LINE deinterlacing), but I felt this would
be harder to implement and even harder to debug in the future.