Fix non-POSIX automake file variable warning
This addresses warnings about the use of the $(wildcard) function.
It has been replaced by a plain shell glob.  There is one case where
this will result in an error, when there are no matches, but that is
when the targets that use it make no sense.  It is probably good to
get an error in that case.