Double quote code passed to AC_LANG_PROGRAM
Upon re-reading the autoconf documentation, this seemed the right thing
to do.  The AC_LANG_PROGRAM macro does not quote its arguments so any
occurence of [ and ] will get stripped out if not double quoted.

Current code snippets passed do not use them but it is better to use a
consistent quoting approach for such snippets.  In other places in the
configure.ac file double quoting is used already.