sanei_pp: fix [-Wcpp] compiler warning
The `#warning` directive is turned into an error by `-Werror`.  The
directive triggers is guaranteed to trigger with __GNUC__ when no
parallel support is available.  This would prevent strict CI builds.

Output a message with all compilers that understand the pragma.  Those
that don't should ignore it, according to C99.