avision.c: Recognize "backlight on" status
Avision scanners can return an additional light-status value meaning
"Transparent Media Adapter back-light is on".  Accept this as
equivalent to "on" during the lamp-warmup check prior to starting a
scan.  Re-check the light status after sending the window command, as
the light selection might have been changed and a further warm-up
may be required.