plustek: Fix  [-Wpedantic] compiler warning
The warning reads:

  enumerator value for '_PS_INP_MIO6' is not an integer constant
  expression

and that comes about as _PS_INP_MIO6 evaluates to 0x8000000, which
does not fit in the *signed* integer range which has been allotted
to enumeration constants by ISO C99.

Seeing that the _PS_INP* values appear to be bitflags (despite the
arithmatic additions in backend/plustek-usbdev.c!) it is safer to
use #defines instead of an enum.