lib/isfdtype.c: Remove badly broken substitute
The substitute completely ignores the fdtype argument and only works
correctly for S_IFSOCK values.  This happened to be the only way the
function was invoked but for safety's sake this has been replaced by
a local implementation that does pretty much the same thing, without
the misleading bit.

Found courtesy of a [-Wunused-parameter] compiler warning on Alpine.