Add JPEG scanning, MB2100/2700 to PIXMA
The Maxify scanners seem to only be able to return JPEG data for ADF
sources.  Attempting to send a gamma LUT will result in an error on
sane_start when trying to use the ADF.  Flatbed scanning is unaffected
and runs fine with LUTs like prior models.

This patch adds support to the PIXMA backend for returned JPEG scans
and keys it off of the new capability, ADF_JPEG.

Tested on a Maxify MB2120 using the ADF and the flatbed scanner.

This may also fix other Maxify MB* models where it seems the ADF does
not with with the prior code, but I only have the MB2120 to test.