Return NO_DOCS after last ADF page, not IO_ERROR
When scanning from ADF in MP150 based scanners, after the last page is
scanned the XML session-close was sent.  However, when using the ADF, if
the frontend called sane_start/sane_read(), it would try and send an
XML command that was not valid after the session abort.  This would give
an IO error up through the stack.

Now, check if ADF scanning is happening and on reads after the last page
return SANE_STATUS_NO_DOCS on read. and abort.

Finally, minor fix of max DPI for Pixma MAXIFY MB21xx/27xx to 1200 DPI.