backend/avision: i1120: set AV_NO_REAR
Kodak broke rear-only scan (bitset3[3]) when they modified the Avision
firmware. It would be possible to add a work-around that drops front data and
writes rear data only. Because of deinterlacing, offsets etc. this is way too
complicated. The simpliest solution to scan rear-only is to turn around the
paper stack and use front scan instead.

Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>