japi: Fix BUILT_SOURCES dependency
Newer automake versions have the dist target depend on BUILT_SOURCES
so that variable should not heavy build dependencies.  This fixes the
dist target sufficiently, for CI purposes at least, in that it leads
to inclusion of the same list of files.  Note that CI does not build
anything in the japi/ directory.