Don't disable backends using pthreads directly.  Fixes #153
The use of pthreads was disabled in case of a non-integral pthread_t
type in 8b9a293fac4ba7796e24de439c30834f0f9a0392 as the sanei_thread
implementation assumes an integral type.  This also disabled building
of the kvs40xx and mustek_usb2 backends which use pthreads directly.

This re-enables the build of these backends.