magicolor: Added security remediation for pixels_per_line.
This implements a security issue reported by GitHub Security Lab.
The details are disclosed in GitLab issue #279.
The issue relates to an invalid scan parameter block being sent to
the backend containing 8 bytes of 0x00 which leads to pixels_per_line
being set to 0. Later arithmetic involves the division by this value
which causes a div by zero crash.