backend/avision: fix lock-up of scanners not supporting fastfeed
fd79655 introduced a regression, where some scanners would lock-up
completely and require a full power cycle.

To fix that, do not call release_unit for any scanners not supporting
fastfeed for now, since that leads to lock-ups on at least one device.
Currently, it isn't clear what exactly is causing the problems, so it
maybe be enabled in a correct way again, later.

Resolves issue !337.

Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>