dell1600n_net: return empty list when local devices are requested
!502 made the dell1600n_net backend return GOOD status when local
devices are requested. This is problematic since in that case
the caller may assume that the passed device list has been populated.
However, this is not the case for this backend, since the device list
is not modified in any way in that case. The caller trying to access
the device list may lead to undesired consequences. Fix that by
returning an empty list if 'local_only' is true.