epson2: implement full range manual focus settings
Allow the whole range of manual focus settings as an expert option. This
is also now independent of a TPU being installed or not. The old default
values are applied if the user changes the source type (either FLATBED
or TPU). Adding a focus group here because autofocus options will come
with later patches.

Signed-off-by: Wolfram Sang <wsa@kernel.org>