test: do not leak initial values of string options
The default values for string options are set using static (i.e.
non-malloced) strings. Later, if new values are loaded from the config
file, those SANE_Strings will then point to dynamically allocated memory
which is eventually leaked.

Change the initial values for the string options to NULL, and initialize
them to the proper values within sane_init() using strdup(). This way,
whenever the value for the string is changed, we can safely free() the
previous value.

This eliminates the other main source of memory leaks in the test backend.

This patch also updates sane_init for the test backend to properly cleanup
memory if it fails.