COPYING: Sync with current plain text version on the FSF site
This updates the Free Software Foundation's postal address.

It replaces references to the Library General Public License (LGPL-2)
with references to the GNU Lesser Public License (LGPL-2.1) as well.
These licenses differ slightly, the latter adds section 6b, but the
references are *outside* the TERMS AND CONDITIONS section.  Moreover,
the first reference is in connection with clarifying the licensing of
Free Software Foundation software and the second reference is merely a
suggestion to use the LGPL in case you want to allow linking your code
with proprietary software.

In view of that, these changes in COPYING do *not* have an effect on
the licensing status of the SANE Backends.

Re #320.