plustek: Adjust gain calculation to avoid calling labs()
labs() has a signed parameter. However the argument to it here was the
difference between two unsigned values, which itself remains unsigned.
GCC warned that using this in labs() might not have the intended result.

By definition though, dwInc >= m_dwIdealGain >= dwDec, so labs() is not
even needed in this expression.