Makefile: Utilize Automake macros for silent rules output
Building with --enable-silent-rules makes it easier to identify errors
or warnings in the build output. Automake provides macros for custom
rules to use, so that they will print a similar message to rules which
run the compiler or linker:

...
  CC       libpieusb_la-pieusb.lo
  CCLD     libpieusb.la
  CCLD     libsane-pieusb.la
  CC       libsane_pixma_la-pixma-s.lo
  GEN      pixma/pixma_sane_options.h
  GEN      pixma/pixma_sane_options.c
  CC       pixma/libpixma_la-pixma.lo
...

This does not change the current output if silent rules are disabled.