tools: Fix inconsistent formatting when generating ChangeLog files
Use 12-digit short hashes (which appear in merge commits). The number
of digits required to unambiguously identify a commit increases as the
Git repository grows. The ChangeLog for the 1.0.27 release has 7-digit
short hashes, which are no longer meaningful: 9 digits are needed now.
Forcing 12 digits to display here is the solution in the Linux kernel.

Do not "decorate" the log with branch or tag names. It is understood
that each file starts at a specific release tag (e.g. 1.0.27) and ends
at the next tag (1.0.28), or at HEAD for development snapshots. Topic
branch names, or the refs "master" and "HEAD", do not need labeling.